Changes to FTPManager, XMLManager and app.py
FTPManager: If there is no new Data to fetch check_for_new_data() now returns an empty list

XMLManager: The XMLManager no longer gets an XML file as Attribute instead the get_data() gets the path of the XML file as attribute

app.py: Made the imports for FTPManager and XMLManager correct. changed the order of the imports so python packages are on top and our packages follow below (for visibility).
Implemented commented code in line 46-51 that needs to be teset.
Added a new TODO for the weekend